# docs/swagger.yaml
basePath: /api/v1
definitions:
  fiber.Map:
    additionalProperties: true
    type: object
  notification-system_internal_notification_model.Notification:
    properties:
      channels:
        description: email, sms, etc.
        items:
          type: string
        type: array
      created_by_id:
        type: string
      created_by_name:
        type: string
      created_date:
        description: |-
          CreatedDate is the time the notification was created
          example: 2025-07-08T14:00:00Z
        type: string
      maximum_retries:
        type: string
      message:
        type: string
      modified_by_id:
        type: string
      modified_by_name:
        type: string
      modified_date:
        type: string
      notification_id:
        description: custom ID
        type: string
      priority:
        type: string
      send_at:
        type: string
      type:
        description: transactional, promotional, etc.
        type: string
      user_id:
        type: string
    required:
    - channels
    - created_by_id
    - created_by_name
    - created_date
    - maximum_retries
    - message
    - modified_by_id
    - modified_by_name
    - modified_date
    - notification_id
    - priority
    - send_at
    - type
    - user_id
    type: object
  notification-system_internal_userpreference_model.Channel:
    properties:
      promotional:
        items:
          type: string
        type: array
      system_alerts:
        items:
          type: string
        type: array
      transactional:
        items:
          type: string
        type: array
    type: object
  notification-system_internal_userpreference_model.DailyLimit:
    properties:
      promotional_limit:
        type: string
      promotional_sent_today:
        type: string
    type: object
  notification-system_internal_userpreference_model.DeliveryTime:
    properties:
      enabled:
        type: boolean
      end_time:
        type: string
      start_time:
        type: string
      time_zone:
        type: string
    type: object
  notification-system_internal_userpreference_model.NoDisturb:
    properties:
      enabled:
        type: boolean
      end_date_time:
        type: string
      start_date_time:
        type: string
      time_zone:
        type: string
    type: object
  notification-system_internal_userpreference_model.Preference:
    properties:
      channels:
        $ref: '#/definitions/notification-system_internal_userpreference_model.Channel'
      daily_limit_details:
        $ref: '#/definitions/notification-system_internal_userpreference_model.DailyLimit'
      delivery_time:
        $ref: '#/definitions/notification-system_internal_userpreference_model.DeliveryTime'
      no_disturb_details:
        $ref: '#/definitions/notification-system_internal_userpreference_model.NoDisturb'
    type: object
  notification-system_internal_userpreference_model.UserPreference:
    properties:
      created_by_id:
        type: string
      created_by_name:
        type: string
      created_date:
        type: string
      modified_by_id:
        type: string
      modified_by_name:
        type: string
      modified_date:
        type: string
      preferences:
        $ref: '#/definitions/notification-system_internal_userpreference_model.Preference'
      user_id:
        type: string
      user_preference_id:
        type: string
    required:
    - created_by_id
    - created_by_name
    - created_date
    - modified_by_id
    - modified_by_name
    - modified_date
    - user_id
    - user_preference_id
    type: object
host: localhost:4001
info:
  contact:
    name: Paul
  description: API docs for Notification System
  title: Notification System API
  version: "1.0"
paths:
  /debug:
    get:
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Dummy check
      tags:
      - debug
  /notifications:
    post:
      consumes:
      - application/json
      description: Create a notification and store it in MongoDB
      parameters:
      - description: Notification payload
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/notification-system_internal_notification_model.Notification'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/notification-system_internal_notification_model.Notification'
      summary: Create a new notification
      tags:
      - notifications
  /ping:
    get:
      description: Just ping
      responses:
        "200":
          description: pong
          schema:
            type: string
      summary: Ping test
      tags:
      - test
  /test:
    get:
      description: Dummy route
      responses:
        "200":
          description: test
          schema:
            type: string
      summary: Test
      tags:
      - test
  /user-preferences:
    post:
      consumes:
      - application/json
      description: Stores user preferences for a given user
      parameters:
      - description: User Preference
        in: body
        name: user_preference
        required: true
        schema:
          $ref: '#/definitions/notification-system_internal_userpreference_model.UserPreference'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Create user preference
      tags:
      - user-preferences
  /user-preferences/{user_id}:
    get:
      description: Fetch user preference by user ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification-system_internal_userpreference_model.UserPreference'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Get user preference
      tags:
      - user-preferences
    put:
      consumes:
      - application/json
      description: Update an existing user preference by ID
      parameters:
      - description: User Preference ID
        in: path
        name: id
        required: true
        type: string
      - description: User Preference Payload
        in: body
        name: userPreference
        required: true
        schema:
          $ref: '#/definitions/notification-system_internal_userpreference_model.UserPreference'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Update user preferences
      tags:
      - user-preferences
swagger: "2.0"